backend/requirements.txt
# Backend Dependencies for Smart Clothes

# FastAPI & Async
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication
PyJWT==2.8.1
bcrypt==4.1.1

# Google Cloud Storage
google-cloud-storage==2.10.0
google-auth==2.25.2

# Image Processing
Pillow==10.1.0
rembg==2.0.50

# AI - Google Services
google-generativeai==0.3.5
google-cloud-aiplatform==1.38.0

# Utilities
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6
